A simpleStorage contract
This is my first time posting about my web3 development journey, specifically on Cyber Security on github, even though i have long had the experience, i think it will be helpful if i start sharing it with the world. I prefer to that with the very basic of the journey, for encouragement and easy way to track my progress.